An attempt at a channel select statement. Seems conceptually sound, but there are strange crashes.
The crashes seem to be occurring in and around the new Select() function, which makes heavy use of objects typed as one of their protocols. Is that a fragile area?

In order to work with Select, channels implement a Selectable protocol. Selectable defines the interface by which the Select function registers another channel, through which it waits for a notification.
This special channel is a singleton which accepts a single write before automatically closing itself. This makes it reasonably immune to contention, which also enables it to carry a data payload across threads on behalf of the Selectable channel that used it.

The switch case that follows the Select call is unfortunately unwieldy. For some reason an object reference is not valid as the control expression for a switch statement. The pattern expression for each case then becomes much more complex, requiring condition (where) clauses.