An unbuffered chan implementation for GCD. Still has a race condition.
The solution is to *always* resume one queue upon suspending the other.
Resuming one queue is not relevant to the other , since if both queues are empty they should both be active.

The reason the buffered version did not have this problem is that the number of situations where a queue get suspended is so much smaller.
The suspend/resume rule applies to all cases, however, and should be enforced.