Sender and Receiver, redone as structs.
With this, the performance of the no-contention message passing for the single-element buffered channel is down to about 165% of the equivalent operation in go (~260ns vs 160ns).

Message passing with contention is much slower than go, but the amount of time beyond the simple thread-swapping delay seems more than expected.