Less C code in PointerQueue, which surprisingly results in quicker (and dirtier) typecasting.
The enqueue() and dequeue() calls now only go directly to the system calls, rather than the C intermediary.
Object creation and destruction (and verification) still use C, but those are not as time-critical.

This version of PointerQueue is about 20% faster than the previous. The repeated memory allocations are and will remain a performance problem.