A successful use of a proper waiting path for selectGet().
For the case where there is no writer already waiting, selectGet() will now get in a waiting mode wherein it signals the writer thread that it too should stop. This enables a 3-way thread rendez-vous, which is the only correct way to do this, as off-putting as it may be.